bsp generation working
No doorways made yet, and maps are very bland, but it is working.  Need to work on spicing it up and creating doorways yet.  May abandon BSP in the end, but this is a good excerise for me.  I'm so bad at algo design that any chance I get to tackle it I need to do it rather than take the easy way out.